Remove generator and overview
