+++ bin/post_compile
@@ -10,10 +10,8 @@ puts-step() {
   echo "-----> $@"
 }
 
-puts-step "Generating images..."
-
-python scripts/generate_sample_images.py 2>&1 | indent
-
 puts-step "Running migrations..."
-
 python manage.py db upgrade 2>&1 | indent
+
+puts-step "Generating images..."
+python scripts/generate_sample_images.py 2>&1 | indent
