+++ memegen/routes/__init__.py
@@ -1,10 +1,10 @@
-from . import static
-from . import root
-from . import templates
-from . import links
-from . import image
-from . import overview
-from . import generator
-from . import latest
+
 from . import aliases
+from . import image
+from . import index
+from . import latest
+from . import links
 from . import magic
+from . import root
+from . import static
+from . import templates
