--- memegen/templates/generator.html
@@ -1,168 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Meme Generator</title>
-  <script src="//code.jquery.com/jquery-2.2.0.min.js"></script>
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.min.css" rel="stylesheet" />
-  <link href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.2-rc.1/css/select2.min.css" rel="stylesheet" />
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.2-rc.1/js/select2.min.js"></script>
-  <style>
-    h1, h2, h3, h4, h5, h6 {
-      color: #159957;
-    }
-
-    body {
-      font-family: "Open Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;
-      color: #606c71;
-    }
-    .select2-dropdown.select2-dropdown--below{
-      width: 448px !important;
-    }
-
-    .select2-container--default .select2-selection--single{
-      padding:6px;
-      height: 60px;
-      width: 448px;
-      font-size: 1.2em;
-      position: relative;
-    }
-    .select2-container--default .select2-selection--single .select2-selection__arrow {
-      color: #fff;
-      font-size: 1.3em;
-      padding: 4px 12px;
-      height: 50px;
-      position: absolute;
-      top: 0px;
-      right: 0px;
-      width: 20px;
-    }
-    .meme-generator {
-      margin-top: 33px;
-    }
-    .meme {
-      height: 50px;
-      line-height: 50px;
-    }
-    .meme img {
-      margin-right: 8px;
-    }
-    .meme-text {
-      vertical-align: top;
-    }
-    .meme-templates {
-      margin-top: 60px;
-    }
-  </style>
-</head>
-<body>
-
-  <div class="container">
-    <div class="row meme-generator">
-      <div class="three columns"></div>
-      <h1>Meme Generator</h1>
-    </div>
-
-    <div class="row">
-      <div class="six columns">
-        <form>
-          <select class="js-meme-selector" name="meme">
-            {% for img in imgs %}
-              <option value="{{ img.key }}" data-url="{{ img.url }}">
-                {{ img.name }}
-              </option>
-            {% endfor %}
-          </select>
-
-          <label for="top">Top Text</label>
-          <input type="text" class="u-full-width" id="meme-text-top" name="top" value="I don't know what this meme is for"/>
-
-          <label for="bottom" >Bottom Text</label>
-          <input type="text" class="u-full-width" id="meme-text-bottom" name="bottom" value="and at this point I'm too afraid to ask"/>
-
-          <button type="button" id="btn-generate" class="button button-primary">Generate</button>
-        </form>
-      </div>
-
-      <div class="six columns">
-        <img width='400' id="meme-image" src="http://memegen.link/afraid/i-don't-know-what-this-meme-is-for/and-at-this-point-i'm-too-afraid-to-ask.jpg" />
-        <div class="">
-          <a id="meme-link" href="http://memegen.link/afraid/i-don't-know-what-this-meme-is-for/and-at-this-point-i'm-too-afraid-to-ask.jpg">http://memegen.link/afraid/i-don't-know-what-this-meme-is-for/and-at-this-point-i'm-too-afraid-to-ask.jpg</a>
-        </div>
-      </div>
-    </div>
-
-    <div class="meme-templates">
-      <h1>Available Templates</h1>
-      {% for img in imgs %}
-        {% if loop.index is divisibleby 3 %}
-        <div class="row">
-        {% endif %}
-          <div class="three columns">
-            <div>
-              <span>{{img.name}}</span>
-              <span>[/{{img.key}}]</span>
-            </div>
-            <img width="240" src="{{ img.url }}" />
-          </div>
-          <div class="one columns">&nbsp;</div>
-        {% if loop.index is divisibleby 3 %}
-        </div>
-        {% endif %}
-      {% endfor %}
-    </div>
-  </div>
-
-
-
-  <script type="text/javascript">
-
-    var memeBase = "http://memegen.link/";
-
-    function escapeMemeText(str) {
-        return str
-          .replace(/-/g, '--')
-          .replace(/_/g, '__')
-          .replace(/\?/g, '~q')
-          .replace(/%/g, '~p')
-          .replace(/\s+/g, '-');
-      }
-
-    function formatMeme(meme) {
-      if (!meme.element) { return meme.text; }
-      var source = encodeURI($(meme.element).attr('data-url')).replace(/,/g, '%2C').replace(/'/g, '%27');
-      var $meme = $(
-        "<div class='meme'><img height='50' src='" + source + "' /><span class='meme-text'>" + meme.text + "</span></div>"
-      );
-      return $meme;
-    };
-
-    function generateMeme() {
-      var memeId = $('.js-meme-selector').val();
-      var memeTextTop = escapeMemeText($('#meme-text-top').val());
-      var memeTextBottom = escapeMemeText($('#meme-text-bottom').val());
-      if (memeId && memeTextTop && memeTextBottom) {
-        var url = memeBase + memeId + '/' + memeTextTop + '/' + memeTextBottom + '.jpg';
-        $("#meme-link").attr('href', url);
-        $("#meme-link").text(url);
-        $("#meme-image").attr('src', encodeURI(url).replace(/,/g, '%2C').replace(/'/g, '%27'));
-      }
-    }
-
-    $('.js-meme-selector').select2({
-      templateResult: formatMeme,
-      templateSelection: formatMeme
-    });
-
-    $( ".js-meme-selector" ).on('change', function() {
-      generateMeme();
-    });
-
-    $("#btn-generate").on('click', function(e) {
-      generateMeme();
-    });
-
-    $( ".js-meme-selector" ).val('afraid').trigger("change");
-  </script>
-
-</body>
-</html>
