--- tests/test_routes_overview.py
@@ -1,12 +1,0 @@
-# pylint: disable=no-self-use
-# pylint: disable=misplaced-comparison-constant
-
-
-class TestOverview:
-
-    def test_get_overview(self, client):
-        response = client.get("/overview")
-
-        assert 200 == response.status_code
-        assert 'text/html' == response.mimetype
-        assert '<img src="' in response.get_data(as_text=True)
